+++ tests/TestCase/Model/EntityTest.php
@@ -203,6 +203,20 @@ class EntityTest extends \PHPUnit\Framework\TestCase
         $this->assertTrue($entity->modified('foo'));
     }
 
+    public function testStates()
+    {
+        $entity = new Entity(['name' => 'test']);
+        $this->assertFalse($entity->created());
+        $this->assertFalse($entity->updated());
+        $this->assertFalse($entity->saved());
+        $this->assertFalse($entity->deleted());
+
+        $this->assertTrue($entity->created(true));
+        $this->assertTrue($entity->updated(true));
+        $this->assertTrue($entity->saved(true));
+        $this->assertTrue($entity->deleted(true));
+    }
+
     public function testToJson()
     {
         $data = [